Added a few things:
Player can now iterate how many checkpoints theyve collected. If it is equal to the number of checkpoints, the player "wins", which is very loosely defined as of now.

Checkpoint now holds a number, which is used to check if checkpoints are visited in the right order.

Board has had numberOfCheckpoints added, which currently has no value but should be set